clean up and document kapi crate to prep
